+++ examples/kyb-app/src/domains/workflows/types.ts
@@ -4,7 +4,6 @@ import {
 } from '@app/domains/workflows/flow-data.type';
 import { AnyObject } from '@ballerine/ui';
 import { RJSFSchema } from '@rjsf/utils';
-import { CreateEndUserDto } from '@app/domains/end-user';
 
 export interface WorkflowUBO {
   entity: {
@@ -14,6 +13,7 @@ export interface WorkflowUBO {
       firstName: string;
       lastName: string;
       email: string;
+      dateOfBirth: string;
       additionalInfo?: Record<PropertyKey, unknown>;
     };
   };
@@ -28,21 +28,53 @@ export interface WorkflowRunDocument {
   }[];
 }
 
+export interface WorkflowEntity {
+  type: 'business';
+  website: string;
+  companyName: string;
+  address: string;
+  country: string;
+  registrationNumber: string;
+  customerCompany: string;
+  email: string;
+  ubos: WorkflowUBO[];
+  birthDate: string;
+  mainRepresentative: PersonalInformationContext;
+  additionalInfo: {
+    company: {
+      vat: string;
+      type: string;
+      incorporationDate: number;
+      industry: string;
+      model: string;
+      estimateAnnualVolume: string;
+      averageTransactionValue: string;
+    };
+    headquarters: {
+      street: string;
+      zip: string;
+      country: string;
+      state: string;
+      city: string;
+      phone: string;
+    };
+    bank: {
+      country: string;
+      name: string;
+      holder: string;
+      accountNumber: string;
+      currencyCode: string;
+    };
+    //WorkflowFlowData
+    __kyb_snapshot: string;
+  };
+}
+
 export interface WorkflowUpdatePayload {
   workflowId: string;
   endUserId: string;
   businessId: string;
-  entity: {
-    type: 'business';
-    website: string;
-    companyName: string;
-    address: string;
-    country: string;
-    registrationNumber: string;
-    customerCompany: string;
-    ubos: WorkflowUBO[];
-    mainRepresentative: PersonalInformationContext;
-  };
+  entity: WorkflowEntity;
   documents: WorkflowRunDocument[];
 }
 
@@ -56,6 +88,15 @@ export interface TRunWorkflowDocument {
   properies: AnyObject;
   version: string;
   issuingVersion: number;
+}
+
+export interface WorkflowAdditionalInformation {
+  firstName: string;
+  lastName: string;
+  phone: string;
+  email: string;
+  dateOfBirth: string;
+  companyName: string;
 }
 
 export interface TRunWorkflowDto {
@@ -82,7 +123,7 @@ export interface TRunWorkflowDto {
         };
         registrationNumber: string;
         additionalInfo?: {
-          mainRepresentative: CreateEndUserDto;
+          mainRepresentative: WorkflowAdditionalInformation;
           ubos: WorkflowUBO[];
         };
       };
@@ -130,9 +171,7 @@ export interface Workflow {
         };
         website: string;
         registrationNumber: string;
-        additionalInfo: {
-          ubos: WorkflowUBO[];
-        };
+        additionalInfo: WorkflowEntity['additionalInfo'];
         companyName: string;
       };
     };
@@ -149,10 +188,14 @@ export interface UpdateWorkflowDto {
 }
 
 export interface GetFlowDataDto {
-  workflowId?: string;
+  email: string;
 }
 
 export interface UpdateFlowDataDto {
   workflowId?: string;
   payload: WorkflowFlowData;
+}
+
+export interface GetActiveWorkflowDto {
+  email: string;
 }