+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ballerine/blocks
 
+## 0.0.7
+
+### Patch Changes
+
+- Update packages
+- Updated dependencies
+  - @ballerine/common@0.5.16
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.5.15
+
 ## 0.0.5
 
 ### Patch Changes
@@ -11,6 +26,7 @@
 
 ### Patch Changes
 
+- added Autocomplete,DropdownInput,Multiselect,DatePicker,Checkbox,PhoneNumber components
 - Upgrade packages
 - Updated dependencies
   - @ballerine/common@0.5.13